requirements.txt
astroid==2.3.3
certifi==2019.11.28
colorama==0.4.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
paho-mqtt==1.5.0
pip==19.3.1
pylint==2.4.4
setuptools==42.0.2.post20191203
six==1.13.0
wheel==0.33.6
wincertstore==0.2
wrapt==1.11.2
